feat: Update service interface to include additional data fields
This commit updates the service interface in the "service.ts" file to include additional data fields such as image, user, category, title, and description. This change allows for more comprehensive and detailed information to be stored and retrieved for each service.

Refactor useGetAllServices to use updated service interface

This commit refactors the useGetAllServices function in the "useGetAllServices.ts" file to use the updated service interface. The services variable is now of type Ref<services[]> instead of Ref<service[]>, ensuring compatibility with the updated data structure.

Update freelancerCard component to use new data fields

This commit updates the freelancerCard component in the "freelancerCard.vue" file to use the new data fields from the updated service interface. The first_name field is now accessed as a prop, and the image field is commented out temporarily.

Update FreelancersView component to use new services data

This commit updates the FreelancersView component in the "FreelancersView.vue" file to use the new services data from the useGetAllServices function. The services variable is now accessed in the v-for loop to display the title of each service.

Update paginadorComponent to use paginatorInfo data

This commit updates the paginadorComponent in the "FreelancersView.vue" file to use the paginatorInfo data from the services variable. The lastPage, count, currentPage, and hasMorePages fields are passed as props to the paginadorComponent.